feat(to_dict): graph deserialization of pointers
