+++ dessia_common/models/__init__.py
@@ -1,2 +1,3 @@
 from .cars import catalog
 from .power_test import simulation1, simulation_list
+
